Move connection to console connection
