+++ src/console/index.ts
@@ -1,1 +1,2 @@
-export * from "./connection";
+export * from "./types";
+export * from "./consoleConnection";
